Добавил страницы для сотрудников

+++ services/EmployerService.php
@@ -4,6 +4,7 @@ namespace app\services;
 
 use app\models\Employer;
 use app\repositories\EmployerRepository;
+use app\services\structures\EmployerData;
 
 class EmployerService
 {
@@ -19,25 +20,43 @@ class EmployerService
         return $this->employerRepository->findAll();
     }
 
-    public function create(string $name, string $surname, string $patronymic, int $gender, int $salary): void
+    public function create(EmployerData $employerData, $departments)
     {
-        $employer = Employer::create($name, $surname, $patronymic, $gender, $salary);
+        $employer = Employer::create(
+            $employerData->name,
+            $employerData->surname,
+            $employerData->patronymic,
+            $employerData->gender,
+            $employerData->salary,
+        );
         $this->employerRepository->add($employer);
-        //@todo добавить проверку связей
+        $this->employerRepository->addRelations($employer->id, $departments);
     }
 
-    public function edit(string $id, string $name, string $surname, string $patronymic, int $salary): void
+    public function edit($id, EmployerData $employerData, $departments)
     {
         $employer = $this->employerRepository->find($id);
-        $employer->edit($name, $surname, $patronymic, $salary);
+        $employer->edit(
+            $employerData->name,
+            $employerData->surname,
+            $employerData->patronymic,
+            $employerData->gender,
+            $employerData->salary
+        );
         $this->employerRepository->save($employer);
-        //@todo добавить сохранение связей
+        $this->employerRepository->addRelations($employer->id, $departments);
     }
 
-    public function delete(string $id): void
+    public function delete(string $id)
     {
         $employer = $this->employerRepository->find($id);
+        $relations = $this->employerRepository->checkRelations($id);
+
+        if (count($relations) > 0) {
+            return false;
+        }
         $this->employerRepository->delete($employer);
-        //@todo добавить проверку связей
+
+        return true;
     }
 }
